requirements.txt
requests>=2.31.0
python-whois>=0.8.0
python-nmap>=0.7.1 